Separate the represented character-progression responsibility (#611)
Eighth delivery of the refactor push. 98 methods move into 3 bounded
submodules under session/progression: skills, talents and reputation. Every
file is under the 1,000-line review budget.

session/mod.rs drops from 42,725 to 41,107 production lines. Across the eight
deliveries the root falls from 75,841 to 41,107 - 46 per cent retired into
navigable owners.

The classifier first produced a glyphs module of two methods and nineteen
lines. That is not a responsibility, so glyph slots were merged into talents,
which is where they belong: the two share the represented talent surface and
are published together.

Methods that mention a skill, talent or reputation but belong to spell, item,
quest, creature, gameobject, pet or loot responsibilities already moved with
those families and are not touched again.

Safety established by comparison: the impl surface is 1,237 methods before and
after, an identical multiset, and the ownership baseline reports 3,761
associated items unchanged. All 17 baseline differences are private to
pub(in crate::session); nothing became public, and no registry or bridge row
changed at all.

Verified: cargo test -p wow-world --lib 3,822 passed and 0 failed,
session-ownership-check --syntax-only PASS, check_architecture.py check and
self-test PASS, fmt and diff hygiene clean. Ceiling tightened.
